Refactor: Use a constructor to store calc state
I struggled to keep track of all the variables
and functions I was writing.

Claude told me it's not great to have a lot
of global variables. It suggested using a
CalculatorState function which gave me the idea
to use a constructor.

This refactoring should make it easier to track
calculator properties and helps me break up the
event handler into smaller scoped functions while
passing in the calculator object to track the
property values.