!Refactor: Separate state handling & display build
I realize there was lack of separation of concerns.

Both the update state function and build display
did a bit of each other's duties. Clearly splitting
allowed me to debug what was happening more cleanly.

This is making it easier to understand what's
happening and remove unnecessary code.
But it's also led to a breaking change where the
refactor has broken the = symbol calling operate
and doesn't work with operators if 0 is the
firstNumber.