Fix: After refactor, operate showed wrong result
After refactoring, I call the update display text
function twice. The first time, it correctly
displays the result of the operation. But the
second time, it updates the userInput.

I fixed this bug by doing a check to only call
update display text function if we're not using
the equal button.